+++ src/Codec/CBOR/Cuddle/CDDL/Resolve.hs
@@ -65,17 +65,20 @@ import Data.Hashable
 #if __GLASGOW_HASKELL__ < 910
 import Data.List (foldl')
 #endif
-import Codec.CBOR.Cuddle.CDDL.CBORGenerator (
+import Codec.CBOR.Cuddle.CDDL.CTreePhase (CTreePhase, XRule (..))
+import Codec.CBOR.Cuddle.CDDL.Custom.Core (RuleTerm)
+import Codec.CBOR.Cuddle.CDDL.Custom.Generator (
   CBORGen,
   GenPhase,
-  TermValidator,
-  ValidatorPhase,
-  WrappedTerm,
   XXCTree (..),
   withLocalGenBindings,
+ )
+import Codec.CBOR.Cuddle.CDDL.Custom.Validator (
+  TermValidator,
+  ValidatorPhase,
+  XXCTree (..),
   withLocalValidateBindings,
  )
-import Codec.CBOR.Cuddle.CDDL.CTreePhase (CTreePhase, XRule (..))
 import Codec.CBOR.Cuddle.IndexMappable (IndexMappable (..))
 import Data.List.NonEmpty qualified as NE
 import Data.Map.Strict qualified as Map
@@ -102,7 +105,7 @@ newtype PartialCTreeRoot i = PartialCTreeRoot (Map.Map Name (ProvidedParameters 
 
 data CDDLMapEntry = CDDLMapEntry
   { cmeProvidedParameters :: ProvidedParameters (TypeOrGroup CTreePhase)
-  , cmeCustomGenerator :: Maybe (CBORGen WrappedTerm)
+  , cmeCustomGenerator :: Maybe (CBORGen RuleTerm)
   , cmeCustomValidator :: Maybe TermValidator
   }
   deriving (Generic)
@@ -164,7 +167,7 @@ type data OrReferenced
 data instance XXCTree OrReferenced
   = -- | Reference to another node with possible generic arguments supplied
     OrRef Name [CTree OrReferenced]
-  | OGenerator (CBORGen WrappedTerm) (CTree OrReferenced)
+  | OGenerator (CBORGen RuleTerm) (CTree OrReferenced)
   | OValidator TermValidator (CTree OrReferenced)
 
 type data OrReferencedSimple
@@ -412,7 +415,7 @@ instance Pretty (XXCTree i) => Pretty (DistRef i) where
 
 data instance XXCTree DistReferenced
   = DRef (DistRef DistReferenced)
-  | DGenerator (CBORGen WrappedTerm) (CTree DistReferenced)
+  | DGenerator (CBORGen RuleTerm) (CTree DistReferenced)
   | DValidator TermValidator (CTree DistReferenced)
 
 type data DistReferencedNoGen
@@ -468,7 +471,7 @@ type data MonoReferenced
 
 data instance XXCTree MonoReferenced
   = MRuleRef Name
-  | MGenerator (CBORGen WrappedTerm) (CTree MonoReferenced)
+  | MGenerator (CBORGen RuleTerm) (CTree MonoReferenced)
   | MValidator TermValidator (CTree MonoReferenced)
 
 type MonoEnv = BindingEnv DistReferenced MonoReferenced