+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 1.7.0.0
 
+* Rename `WrappedTerm` to `RuleTerm`
+* Rename `RuleTerm` constructors:
+  * `S` -> `SingleTerm`
+  * `G` -> `GroupTerm`
+  * `P` -> `PairTerm`
+* Split `Codec.CBOR.Cuddle.CDDL.CBORGenerator` into:
+  * `Codec.CBOR.Cuddle.CDDL.Custom.Generator`
+  * `Codec.CBOR.Cuddle.CDDL.Custom.Validator`
+  * `Codec.CBOR.Cuddle.CDDL.Custom.Core`
 * Fix `--no-twiddle` argument
 * Add autocompletion hints
 * Add footer, improve output format failure message